some optimizations
* combine qkv into single linear layer to process and split after

* removed need for torch for multinomial logit sampling

* restrict context to blocksize for generation iterations, purely for speed

* added file output writing and running print of next token generation

* tweak hyperparams. still slow as hell but doable on lamda labs A10 (24GB pci) +-4 hours

+++ examples/gpt.py
@@ -7,8 +7,6 @@ import tinygrad.nn as nn
 from tinygrad.nn import optim
 from tinygrad.state import get_parameters
 from tinygrad.tensor import Tensor, dtypes
-import torch # remove if faster multimonial selection is found
-from torch import multinomial
 
 # tiny shakespear input text
 input_url = "https://raw.githubusercontent.com/karpathy/char-rnn/master/data/tinyshakespeare/input.txt"
@@ -16,20 +14,27 @@ input_url = "https://raw.githubusercontent.com/karpathy/char-rnn/master/data/tin
 Tensor.manual_seed(31337)
 
 # hyperparameters
-batch_size = 32 # how many independent sequences will we process in parallel?
-block_size = 8 # what is the maximum context length for predictions?
-max_iters = 1000
-eval_interval = 100
-learning_rate = 1e-3
-eval_iters = 50
-n_embd = 32
+batch_size = 64 # how many independent sequences will we process in parallel?
+block_size = 128 # what is the maximum context length for predictions?
+max_iters = 5000
+eval_interval = 500
+learning_rate = 3e-4
+eval_iters = 200
+n_embd = 192
 n_head = 4
-n_layer = 3
+n_layer = 4
 dropout = 0.2
 
+BOLD = '\033[1m'
+END = '\033[0m'
+
 def get_input_text(url, file_path, clean):
-  """ fetch and return tiny Shakespear input text """
-  urllib.request.urlretrieve(url, file_path)
+  """ fetch and return tiny Shakespeare input text """
+  if not os.path.exists(file_path):
+    urllib.request.urlretrieve(url, file_path)
+    print("stored tiny shakespeare to " + BOLD + file_path + END)
+  else:
+    print("reading shakespeare from " + BOLD + file_path + END)
   with open(file_path, 'r') as f:
     text = f.read()
   if clean:
@@ -39,7 +44,7 @@ def get_input_text(url, file_path, clean):
 def cross_entropy(out, Y):
   """ negative Log Loss function """
   num_classes = out.shape[-1]
-  YY = Y.flatten().astype(np.int32)
+  YY = Y.numpy().flatten().astype(np.int32)
   y = np.zeros((YY.shape[0], num_classes), np.float32)
   y[range(y.shape[0]),YY] = -1.0*num_classes
   y = y.reshape(list(Y.shape)+[num_classes])
@@ -73,23 +78,28 @@ class Head():
   """ one head of self-attention """
 
   def __init__(self, head_size):
-    self.key = nn.Linear(n_embd, head_size, bias=False)
-    self.query = nn.Linear(n_embd, head_size, bias=False) # usually bias term is added on q
-    self.value = nn.Linear(n_embd, head_size, bias=False) # usually bias term is added on v
+    # optimized k,q,v layer to compute with a single linear transformation
+    self.to_kqv = nn.Linear(n_embd, 3 * head_size, bias=False)
 
   def __call__(self, x):
-    B,T,C = x.shape
-    k = self.key(x)   # (B,T,C)
-    q = self.query(x) # (B,T,C)
+    B, T, _ = x.shape
+    kqv = self.to_kqv(x) # shape: (B, T, 3 * head_size)
+    kqv = kqv.reshape(B, T, 3, -1) # shape: (B, T, 3, head_size)
+
+    # splitting k,q,v
+    k = kqv[:,:,0,:]
+    q = kqv[:,:,1,:]
+    v = kqv[:,:,2,:]
+
     # compute attention scores ("affinities")
-    wei = q @ Tensor.transpose(k, -2, -1) * C**-0.5 # (B, T, hs) @ (B, hs, T) -> (B, T, T)
+    wei = q @ Tensor.transpose(k, -2, -1) * n_embd**-0.5
     # equal to a lower triangular matrix (tril) masked_fill in pytorch
     mask = Tensor(np.triu(np.ones((T,T), dtype=np.float32) * -np.inf, k=1))
     wei = wei + mask
     wei = wei.softmax(-1)
     wei = wei.dropout(dropout)
+
     # perform the weighted aggregation of the values
-    v = self.value(x) # (B,T,hs)
     out = wei @ v  # (B, T, T) @ (B, T, hs) -> (B, T, hs)
     return out
 
@@ -149,14 +159,14 @@ class GPTLanguageModel():
     self.lm_head = nn.Linear(n_embd, vocab_size)
 
   def __call__(self, idx, targets=None):
-    B, T = idx.shape
+    _, T = idx.shape
 
     # idx and targets are both (B,T) tensor of integers
     tok_emb = self.token_embedding_table(idx) # (B,T,C)
     pos_emb = self.position_embedding_table(Tensor.arange(T, dtype=dtypes.int8).reshape(1,T)) # (T,C)
     x = tok_emb + pos_emb # (B,T,C)
-    for block in self.blocks: x = block(x)
-    x = self.ln_f(x)
+    for block in self.blocks: x = block(x) # (B,T,C)
+    x = self.ln_f(x) # (B,T,C)
     logits = self.lm_head(x) # (B,T,vocab_size)
     
     if targets is None:
@@ -164,7 +174,7 @@ class GPTLanguageModel():
     else:
       # log softmax to get predictions for cross_entropy loss calculation
       predictions = logits.log_softmax(-1)
-      loss = cross_entropy(predictions, targets.numpy())
+      loss = cross_entropy(predictions, targets)
     return logits, loss
 
   def generate(self, idx, max_new_tokens):
@@ -179,13 +189,15 @@ class GPTLanguageModel():
       # apply softmax to get probabilities
       probs = logits.softmax(-1) # (B, C)
       # sample from the distribution
-      idx_next = multinomial(torch.tensor(probs.numpy()), num_samples=1).numpy() # (B, 1)
-      # multinomial sampling using numpy  **slow**
-      #idx_next = np.array([np.random.choice(len(p), size=1, p=p) for p in probs.numpy()])
+      idx_next = [np.random.choice(len(p), size=1, p=p) for p in probs.numpy()]
+      #idx_next = multinomial(torch.tensor(probs.numpy()), num_samples=1).numpy() # (B, 1)
+      print(decode(idx_next[0]), flush=True, end='')
       idx_next = Tensor(idx_next)
       # append sampled index to the running sequence
       idx = Tensor.cat(idx, idx_next, dim=1) # (B, T+1)
-      print(decode(idx_next.numpy()[0]), end='')
+      # limit idx to block_size to reduce context length for efficiency
+      # comment this out to use full concatenated idx context in each iteration **slow**
+      idx = idx[:,-block_size:]
     return idx
 
 
@@ -200,13 +212,17 @@ if __name__ == "__main__":
   """
 
   parser = argparse.ArgumentParser(description="""Tiny Shakespeare GPT""")
-  parser.add_argument('--path', default=os.path.join(os.path.sep, 'tmp', 'input.txt'), 
-                      help='Location to save the input text, defaults to /tmp/input.txt')
+  parser.add_argument('--input', default=os.path.join(os.path.sep, 'tmp', 'input.txt'), 
+                      help="Where to save the input text, defaults to '/tmp/input.txt'",
+                      metavar="PATH")
+  parser.add_argument('--output', default=None, 
+                      help='Save the output text, defaults to NO OUTPUT.',
+                      metavar="PATH")
   parser.add_argument('--clean', action="store_true",
                       help='Delete the input text file after run, defaults to False.')
   args = parser.parse_args()
 
-  text = get_input_text(input_url, args.path, args.clean)
+  text = get_input_text(input_url, args.input, args.clean)
 
   # here are all the unique characters that occur in this text
   chars = sorted(list(set(text)))
@@ -229,9 +245,10 @@ if __name__ == "__main__":
   model = GPTLanguageModel(vocab_size)
   
   parameters = get_parameters(model)
-  optimizer = optim.AdamW(parameters, lr=1e-3)
+  optimizer = optim.AdamW(parameters, lr=learning_rate)
 
   Tensor.training = True
+  print("imbuing...")
   for iter in range(max_iters):
 
     # every once in a while evaluate the loss on train and val sets
@@ -245,12 +262,16 @@ if __name__ == "__main__":
     # evaluate the loss
     _, loss = model(xb, yb)
     optimizer.zero_grad()
-    loss.backward()        
+    loss.backward()
     optimizer.step()
   
-  print("final loss: {0}".format(loss.numpy()))
-
   Tensor.training = False
-  context = Tensor.zeros((1, 1), dtypes.int64)
   print("-- hark! --")
-  model.generate(context, max_new_tokens=500).numpy()[0]
+  # generate output tokens
+  context = Tensor.zeros((1, 1), dtypes.int64)
+  out = decode(model.generate(context, max_new_tokens=2000).numpy()[0])
+  print("\n-- exeunt. --")
+  if (args.output is not None):
+    with open(args.output, 'w') as f:
+      print(out, file=f)
+    print("output saved to " + BOLD + args.output + END)
